Logic renamed to lib to follow convention
