--- .github/workflows/tests.yml
@@ -1,19 +1,0 @@
-name: Tests
-
-on: pull_request
-
-jobs:
-  rspec:
-    name: RSpec
-    runs-on: ubuntu-18.04
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-ruby@v1
-        with:
-          ruby-version: 2.6.x
-      - name: Setup RSpec
-        run: |
-          [ -f Gemfile ] && bundle --deployment
-          gem install --no-document rspec:'~>3.0'
-      - name: RSpec Report
-        run: rspec --force-color --format documentation
